+++ app/controllers/manager/tasks_controller.js
@@ -1,5 +1,5 @@
-const Controller = require('./controller');
-const models = require('../models');
+const Controller = require('../controller');
+const models = require('../../models');
 
 class ExamplesController extends Controller {
   // GET /create
@@ -7,7 +7,7 @@ class ExamplesController extends Controller {
     const task = models.Task.build({});
     const team = await this._team(req);
     const memberUsers = await team.getMemberUsers();
-    res.render(`tasks/create`, { task, team, memberUsers });
+    res.render(`manager/tasks/create`, { task, team, memberUsers });
   }
 
   // POST /
@@ -34,7 +34,7 @@ class ExamplesController extends Controller {
 
     await task.save();
     await req.flash('info', `タスク[${task.title}]を保存しました`);
-    res.redirect(`/teams/${team.id}`);
+    res.redirect(`/manager/teams/${team.id}`);
   }
 
   // GET /:id/edit
@@ -42,7 +42,7 @@ class ExamplesController extends Controller {
     const task = await this._task(req);
     const team = await this._team(req);
     const memberUsers = await team.getMemberUsers();
-    res.render('tasks/edit', { task, memberUsers });
+    res.render('manager/tasks/edit', { task, memberUsers });
   }
 
   // PUT or PATCH /:id
@@ -68,7 +68,7 @@ class ExamplesController extends Controller {
     }
     await task.save();
     await req.flash('info', `[${task.title}]を更新しました`);
-    res.redirect(`/teams/${task.teamId}`);
+    res.redirect(`/manager/teams/${task.teamId}`);
   }
 
   async _team(req) {
@@ -91,16 +91,13 @@ class ExamplesController extends Controller {
     if(req.body.displayName) {
       const user = await models.User.findOne({ where : { displayName: req.body.displayName } });
       if (!user) {
-        throw new Error('Task not find');
+        throw new Error('User not find');
       }
       return user;
     } else {
       return null;
     }
   }
-
 }
-
-
 
 module.exports = ExamplesController;
