+++ app/controllers/teams_controller.js
@@ -2,12 +2,6 @@ const Controller = require('./controller');
 const models = require('../models');
 
 class TeamsController extends Controller {
-  // GET /
-  // async index(req, res) {
-  //   // res.render('teams/index', { teams: teams });
-  //   res.render('teams/index', { team: req.body });
-  // }
-
   // GET /create
   async create(_req, res) {
     const team = models.Team.build({});
@@ -22,46 +16,7 @@ class TeamsController extends Controller {
     });
     await team.save();
     await req.flash('info', `新規チーム[${team.name}]を作成しました`);
-    res.redirect(`/teams/${team.id}`);
-  }
-
-  // GET /:id
-  async show(req, res) {
-    const team = await this._team(req);
-    const tasks = await team.getTasks({ 
-      include: ['assignee'],
-      order: [['id', 'DESC']]
-    });
-    res.render('teams/show', { team, tasks });
-  }
-
-  // GET /:id/edit
-  async edit(req, res) {
-    const team = await this._team(req);
-    res.render('teams/edit', { team });
-  }
-
-  // PUT or PATCH /:id
-  async update(req, res) {
-    const team = await this._team(req);
-    team.set(req.body);
-    await team.save();
-    await req.flash('info', `[${team.name}]に更新しました`);
-    res.redirect(`/teams/${team.id}/edit`);
-  }
-
-  // DELETE /:id
-  // async destroy(req, res) {
-  //   await req.flash('info', '削除しました（未実装）');
-  //   res.redirect('/teams/');
-  // }
-
-  async _team(req) {
-    const team = await models.Team.findByPk(req.params.team);
-    if (!team) {
-      throw new Error('Team not find');
-    }
-    return team;
+    res.redirect(`/manager/teams/${team.id}`);
   }
 }
