+++ views/index.pug
@@ -3,7 +3,3 @@ extends layout
 block content
 
   a(href="/teams/create") チーム作成
-//-
-  p Welcome to #{title}
-  p 
-    a(href="/examples") CRUDのリソースのサンプル
