+++ views/manager/tasks/_form.pug
