+++ views/manager/tasks/create.pug
@@ -1,0 +1,7 @@
+extends ../../layout
+
+include ./_form
+
+block content 
+  h1 タスクの新規作成
+  +task_form(`/manager/teams/${team.id}/tasks`, 'post', task, memberUsers, err)
