+++ views/manager/teams/show.pug
@@ -1,12 +1,12 @@
-extends ../layout
+extends ../../layout
 
 block content 
   h1 #{team.name}
-  a(href=`/teams/${team.id}/tasks/create`) 新規タスクの作成
+  a(href=`/manager/teams/${team.id}/tasks/create`) 新規タスクの作成
   br
-  a(href=`/teams/${team.id}/members`) メンバー管理
+  a(href=`/manager/teams/${team.id}/members`) メンバー管理
   br
-  a(href=`/teams/${team.id}/edit`) 編集
+  a(href=`/manager/teams/${team.id}/edit`) 編集
 
   h2 タスク一覧
   - var hasTasks = tasks.length > 0
@@ -30,6 +30,6 @@ block content
               td なし
             td= helpers.formatDateTime(task.createdAt)
             td
-              a(href=`/teams/${team.id}/tasks/${task.id}/edit`) 編集
+              a(href=`/manager/teams/${team.id}/tasks/${task.id}/edit`) 編集
   else
     p タスクはありません
