--- views/tasks/edit.pug
@@ -1,7 +1,0 @@
-extends ../layout
-
-include ./_form
-
-block content 
-  h1 タスクの編集
-  +task_form(`/teams/${task.teamId}/tasks/${task.id}`, 'put', task, memberUsers, err)
